feat<Path>: use matlab generate path

+++ src/main/java/ink/verge/logistics/entity/trans/PathNoId.java
@@ -9,5 +9,5 @@ import lombok.Data;
  */
 @Data
 public class PathNoId {
-    int[] path;
+    double[] path;
 }
